* develop: (36 commits)
  Wed Jun  8 01:30:08 WEST 2016
  Added EndlessRecyclerOnScrollListener.java. Added Endless to Search and ListApps.
  Added OffsetInterface.java to allow easier endless integration.
  Updated com.android.tools.build to 2.1.2.
  Finished search pager logic.
  Added SearchWidget.java bindings.
  Added search suggestions.
  Added GlobalSearchFragment.java.
  Changed realm primaryKeys.
  Tue Jun  7 01:29:48 WEST 2016
  fixed app bricks not showing properly
  minor code refactoring
  added toolbar to settings view with integrated back button
  Mon Jun  6 01:29:41 WEST 2016
  Sun Jun  5 01:29:33 WEST 2016
  Sat Jun  4 01:29:42 WEST 2016
  Fri Jun  3 01:29:32 WEST 2016
  * changes in the db migration script * fixed incorrect layout properties in the app view fragment
  Thu Jun  2 01:29:36 WEST 2016
  Added trusted flag to ListSearchAppsRequest.java.
  ...

# Conflicts:
#	v8engine/src/main/java/cm/aptoide/pt/v8engine/view/recycler/displayable/DisplayablesFactory.java

+++ dataprovider/src/main/java/cm/aptoide/pt/dataprovider/ws/v7/GetAppRequest.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 2016.
- * Modified by Neurophobic Animal on 12/05/2016.
+ * Modified by Neurophobic Animal on 07/06/2016.
  */
 
 package cm.aptoide.pt.dataprovider.ws.v7;
@@ -26,11 +26,11 @@ public class GetAppRequest extends V7<GetApp, GetAppRequest.Body> {
 		super(bypassCache, new Body());
 	}
 
-	public static GetAppRequest of(int appId) {
+	public static GetAppRequest of(long appId) {
 		return of(appId, false);
 	}
 
-	public static GetAppRequest of(int appId, boolean bypassCache) {
+	public static GetAppRequest of(long appId, boolean bypassCache) {
 		GetAppRequest getAppRequest = new GetAppRequest(bypassCache);
 
 		getAppRequest.body.appId = appId;
@@ -58,7 +58,7 @@ public class GetAppRequest extends V7<GetApp, GetAppRequest.Body> {
 
 		private Integer apkId;
 		private String apkMd5sum;
-		private Integer appId;
+		private Long appId;
 		private String lang = Api.LANG;
 		private Integer limit;
 		private List<AppNodes> nodes;
@@ -67,7 +67,7 @@ public class GetAppRequest extends V7<GetApp, GetAppRequest.Body> {
 		private String packageName;
 		private String q = Api.Q;
 		private Integer storeId;
-		private List<Integer> storeIds;
+		private List<Long> storeIds;
 		private String storeName;
 		//		Doesn't make sense without stores_auth_map
 //		private List<String> storeNames;
